requirements.txt
dataclasses==0.6
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.1
djongo==1.2.33
dnspython==1.16.0
gunicorn==19.9.0
pymongo==3.8.0
pytz==2019.1
sqlparse==0.2.4
djangorestframework-jwt==1.11.0
